+++ src/main/java/net/dehydration/item/WaterBowlItem.java
@@ -1,9 +1,12 @@
 package net.dehydration.item;
 
+import java.util.Optional;
+
 import net.dehydration.DehydrationMain;
 import net.dehydration.access.ThirstManagerAccess;
 import net.dehydration.init.ConfigInit;
 import net.dehydration.init.EffectInit;
+import net.dehydration.misc.ThirstTooltipData;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.player.PlayerEntity;
@@ -11,6 +14,7 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ItemUsage;
 import net.minecraft.item.Items;
+import net.minecraft.item.tooltip.TooltipData;
 import net.minecraft.util.Hand;
 import net.minecraft.util.TypedActionResult;
 import net.minecraft.util.UseAction;
@@ -29,7 +33,7 @@ public class WaterBowlItem extends Item {
     public ItemStack finishUsing(ItemStack stack, World world, LivingEntity user) {
         ItemStack itemStack = super.finishUsing(stack, world, user);
 
-        if (user instanceof PlayerEntity) {
+        if (user instanceof PlayerEntity playerEntity) {
             int thirstQuench = 0;
             for (int i = 0; i < DehydrationMain.HYDRATION_TEMPLATES.size(); i++) {
                 if (DehydrationMain.HYDRATION_TEMPLATES.get(i).containsItem(stack.getItem())) {
@@ -42,11 +46,11 @@ public class WaterBowlItem extends Item {
             }
             ((ThirstManagerAccess) user).getThirstManager().add(thirstQuench);
 
-            if (!world.isClient() && this.hasThirstChance && world.random.nextFloat() >= ConfigInit.CONFIG.water_bowl_thirst_chance) {
+            if (!world.isClient() && this.hasThirstChance && world.getRandom().nextFloat() >= ConfigInit.CONFIG.water_bowl_thirst_chance) {
                 user.addStatusEffect(new StatusEffectInstance(EffectInit.THIRST, ConfigInit.CONFIG.potion_bad_thirst_duration / 2, 0, false, false, true));
             }
 
-            if (((PlayerEntity) user).isCreative()) {
+            if (playerEntity.isCreative()) {
                 return itemStack;
             }
         }
@@ -66,6 +70,25 @@ public class WaterBowlItem extends Item {
     @Override
     public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
         return ItemUsage.consumeHeldItem(world, user, hand);
+    }
+
+    @Override
+    public Optional<TooltipData> getTooltipData(ItemStack stack) {
+        if (ConfigInit.CONFIG.thirst_preview) {
+            int thirstQuench = 0;
+            for (int i = 0; i < DehydrationMain.HYDRATION_TEMPLATES.size(); i++) {
+                if (DehydrationMain.HYDRATION_TEMPLATES.get(i).containsItem(stack.getItem())) {
+                    thirstQuench = DehydrationMain.HYDRATION_TEMPLATES.get(i).getHydration();
+                    break;
+                }
+            }
+            if (thirstQuench == 0) {
+                thirstQuench = ConfigInit.CONFIG.water_bowl_quench;
+            }
+            return Optional.of(new ThirstTooltipData(this.hasThirstChance ? 2 : 0, thirstQuench));
+        } else {
+            return super.getTooltipData(stack);
+        }
     }
 
 }