+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 - 
 ### Fixed:
 - Fixed potion crash
+- Fixed bowl tooltip
+- Fixed potion quench
 ### Changed:
 -